requirements.txt
httpx>=0.27.0
pinecone-client>=3.0.0
openai>=1.30.0
